Correcciones en logeo. Se agrego filtros segun rol

+++ Controllers/PresupuestoController.cs
@@ -17,6 +17,7 @@ public class PresupuestoController : Controller
     [HttpGet]
     public ActionResult<List<Presupuestos>> ListarPresupuestos()
     {
+        ViewData["AccessLevels"] = HttpContext.Session.GetString("AccessLevels");
         var lista = _presupuestoRepository.listarPresupuestos();
         return View(lista);
     }
@@ -29,35 +30,55 @@ public class PresupuestoController : Controller
     [HttpGet]
     public IActionResult CrearPresupuesto(int idPresupuesto)
     {
-        var clientes = _clienteRepository.listarClientes();
-        var presupuesto = new Presupuestos(idPresupuesto);
-        var presupuestoViewModel = new CrearPresupuestoViewModel(presupuesto,clientes);
-        return View(presupuestoViewModel);
+        if (HttpContext.Session.GetString("AccessLevels") == "Administrador")
+        {
+            var clientes = _clienteRepository.listarClientes();
+            var presupuesto = new Presupuestos(idPresupuesto);
+            var presupuestoViewModel = new CrearPresupuestoViewModel(presupuesto, clientes);
+            return View(presupuestoViewModel);
+        }
+        return RedirectToAction("Index", "Home");
     }
     [HttpPost]
-    public IActionResult CrearPresupuesto(Presupuestos presupuesto,int idCliente)
+    public IActionResult CrearPresupuesto(Presupuestos presupuesto, int idCliente)
     {
-        _presupuestoRepository.CrearPresupuesto(presupuesto,idCliente);
-        return RedirectToAction("ListarPresupuestos");
+        if (HttpContext.Session.GetString("AccessLevels") == "Administrador")
+        {
+            _presupuestoRepository.CrearPresupuesto(presupuesto, idCliente);
+            return RedirectToAction("ListarPresupuestos");
+        }
+        return RedirectToAction("Index", "Home");
     }
     [HttpGet]
     public IActionResult EliminarPresupuesto(int idPresupuesto)
     {
+        if (HttpContext.Session.GetString("AccessLevels") == "Administrador")
+        {
         _presupuestoRepository.eliminarPresupuesto(idPresupuesto);
         return RedirectToAction("ListarPresupuestos");
+        }
+        return RedirectToAction("Index", "Home");
     }
     [HttpGet]
     public IActionResult AgregarProducto(int idPresupuesto)
     {
+        if (HttpContext.Session.GetString("AccessLevels") == "Administrador")
+        {
         var listaProductos = _productoRepository.listarProductos();
-        var agregarProducto = new AgregarProductoViewModel(idPresupuesto,listaProductos);
+        var agregarProducto = new AgregarProductoViewModel(idPresupuesto, listaProductos);
         return View(agregarProducto);
+        }
+        return RedirectToAction("Index", "Home");
     }
     [HttpPost]
     public IActionResult AgregarProducto(int idPresupuesto, int idProducto, int cantidad)
     {
-        _presupuestoRepository.agregarProducto(idPresupuesto,idProducto,cantidad);
-        return RedirectToAction("VerDetalle",new{idPresupuesto = idPresupuesto});
+        if (HttpContext.Session.GetString("AccessLevels") == "Administrador")
+        {
+        _presupuestoRepository.agregarProducto(idPresupuesto, idProducto, cantidad);
+        return RedirectToAction("VerDetalle", new { idPresupuesto = idPresupuesto });
+        }
+        return RedirectToAction("Index", "Home");
     }
 
 }
